minor: refactor schema restrictions into their own interface so there can be multiple implementations
