+++ src/Restrictions/IFormAccessRestriction.cs
@@ -1,0 +1,9 @@
+using form_builder.Models;
+
+namespace form_builder.Restrictions
+{
+    public interface IFormAccessRestriction
+    {
+        bool IsRestricted(FormSchema baseForm);
+    }
+}
